requirements.txt
absl-py==2.3.0
addict==2.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.6.3
asttokens==3.0.0
astunparse==1.6.3
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
basicsr==1.4.2
beautifulsoup4==4.13.4
botocore==1.37.38
Brotli==1.1.0
cachetools==4.2.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==7.1.2
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
facexlib==0.3.0
ffmpeg-python==0.2.0
filelock==3.16.1
filterpy==1.4.5
flatbuffers==1.12
fonttools==4.57.0
fsspec==2023.6.0
future==1.0.0
gast==0.4.0
gdown==4.7.1
gfpgan==1.3.8
gitdb==4.0.12
GitPython==3.1.42
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
gTTS==2.5.4
h11==0.16.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.23.5
hyperframe==6.0.1
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.1
jmespath==1.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.41.1
lmdb==1.6.2
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
moviepy==1.0.3
mpmath==1.3.0
mutagen==1.47.0
networkx==3.1
norbert==0.2.1
numba==0.58.1
numpy==1.23.1
oauthlib==3.3.0
onnx==1.17.0
openai==1.93.0
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==23.2
pandas==1.5.3
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==4.3.6
playwright==1.48.0
pluggy==1.5.0
prettytable==3.9.0
proglog==0.1.12
prompt_toolkit==3.0.51
protobuf==3.19.6
pure_eval==0.2.3
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.0.0
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2025.2
PyWavelets==1.4.1
pywin32-ctypes==0.2.3
PyYAML==6.0.1
qai-hub==0.30.0
qai-hub-models==0.16.2
realesrgan==0.3.0
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9.1
s3transfer==0.10.4
schema==0.7.5
scikit-image==0.21.0
scipy==1.8.1
seaborn==0.11.0
semver==3.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
spleeter==2.4.0
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tb-nightly==2.14.0a20230808
Telethon==1.40.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tifffile==2023.7.10
tomli==2.2.1
torch==2.1.2
torchvision==0.16.2
tornado==6.1
tqdm==4.66.2
traitlets==5.14.3
typer==0.3.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.2
urllib3==1.26.18
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.6
wrapt==1.17.2
yapf==0.43.0
yt-dlp==2024.10.22
zipp==3.20.2
